feat: std.io.fs — stat, walk_dir, glob + supporting compiler/stdlib fixes
New stdlib API:
- stat(path) / FileInfo { kind, size } — follows symlinks.
  exists / is_dir / is_file are thin wrappers.
- walk_dir(root, allocator?) -> WalkIter. DFS pre-order. Path is a view
  into iterator-owned StringBuilder, invalidated on next next().
Symlinks
  are not followed (cycle safety). Errors surface via walk.err() after
  iteration completes.
- glob(pattern, allocator?) -> GlobIter, match_glob(pattern, path).
  Built on walk_dir. Supports '*' (segment-local), '**' (cross-segment),
  '?' (single non-separator byte). Character classes not yet
implemented.
  Literal prefix is used as the walk root so patterns like src/**/*.f do
  not enumerate the whole filesystem.

Supporting fixes:

- HmAstLowering: for-loop lowering now handles iter() returning a
  reference type. Previously it always alloca'd a slot and stored the
  iter result, producing T** where next() expected T*. WalkIter needs
  iter(&WalkIter) -> &WalkIter because copy-by-value would alias the
  iterator's List/StringBuilder heap pointers.
- List.op_index on &List returns &T (was T) so callers can mutate
  elements in place via list[i].field = ... without reaching into
  list.ptr pointer arithmetic. Value-form op_index(List, usize) -> T
  is unchanged for by-value reads.
- StringBuilder.ensure_capacity: had an unsigned-underflow bug when
  requested capacity was already satisfied; rewrote to compute the
  additional reserve correctly.
- core.rtti.TypeInfo: size/align were u8, capping at 255 bytes.
  Promoted to usize so structs like DirIter (288 bytes) compile.
- fs.c stat: Windows branch was reading GetLastError() for errors from
  _stat64, which is a CRT function and reports via errno. Hoisted the
  errno→FsError mapper above the platform split so both branches use it.

Tests: fs/stat_basic.f, fs/walk_dir_basic.f, fs/glob_basic.f.
list_basic.f updated for the new op_index(&List) reference return.

+++ stdlib/core/rtti.f
@@ -23,8 +23,8 @@ pub type ParamInfo = struct {
 
 pub type TypeInfo = struct {
     name: String
-    size: u8
-    align: u8
+    size: usize
+    align: usize
     kind: TypeKind
     type_params: String[]
     type_args: &TypeInfo[]
